ai_exam_oracle/frontend/AI Exam Paper Generator/backend/requirements.txt
fastapi==0.115.12
gunicorn==23.0.0
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.6
python-multipart==0.0.18
openai==1.59.7
pypdf2==3.0.1
python-docx==1.1.2
chromadb==0.6.4
sentence-transformers==3.4.1
requests==2.32.3
pymysql==1.1.1
python-dotenv==1.0.1
aiofiles==24.1.0
ollama==0.4.7
wikipedia-api==0.7.2
datasets==3.2.0
huggingface_hub==0.27.1
